Introduce CNAME monitoring feature
This change introduces the feature that octo-proxy can close connections
when it detects a failover based on a change to a CNAME record.

It is occasionally desirable to failover traffic from one target to
another. A popular way of doing this failover is by changing a CNAME.
The next connection that is made will use the updated DNS records to
resolve so new connections will me made to the newly desired targets.
Existing connections will remain intact unless otherwise disconnected;
this behaviour may or may not be desirable depending on the application.